+++ src/app/modules/user/user-projects/user-projects.component.html
@@ -16,26 +16,7 @@
    If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
 -->
 
-<!--
-
-  Digital Excellence Copyright (C) 2020 Brend Smits
-
-   This program is free software: you can redistribute it and/or modify
-   it under the terms of the GNU Lesser General Public License as published
-   by the Free Software Foundation version 3 of the License.
-
-   This program is distributed in the hope that it will be useful,
-   but WITHOUT ANY WARRANTY; without even the implied warranty
-   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-   See the GNU Lesser General Public License for more details.
-
-   You can find a copy of the GNU Lesser General Public License
-   along with this program, in the LICENSE.md file in the root project directory.
-   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
-
- -->
-
- <div class="row">
+<div class="row">
 
 
   <div class="col-md projects-overview-wrapper">
@@ -65,7 +46,7 @@
     <div class="userinfo">
       <img alt="" class="profile-picture" loading="lazy" src="assets/images/profile-placeholder.png" />
       <div class="usercreds">
-       <h3> {{ name }} </h3>
+       <h3> {{ userName }} </h3>
        <div class="user-amounts">
        <div class="grid-item"><p>Projects:</p></div>
        <div class="grid-item"><p>{{ userprojects.length }}</p></div>
@@ -73,8 +54,6 @@
        </div>
       </div>
     </div>
-
-  
 
     <ng-container *ngIf="!projectsEmpty(); else noProjects">
       <div class="project-list-wrapper" [ngClass]="this.showListView ? 'list' : 'grid'">
@@ -92,7 +71,5 @@
       <h2 class="fas fa-spinner fa-pulse loading-circle" *ngIf="projectsLoading"></h2>
       <p *ngIf="!projectsLoading">There are no projects available.</p>
     </ng-template>
-
-
   </div>
 </div>
