+++ src/app/modules/user/user.module.ts
@@ -1,26 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
 import { UserRoutingModule } from './user-routing.module';
-import { UserProjectComponent } from './user-project/user-project.component';
-import { PaginationModule } from 'ngx-bootstrap/pagination';
-import { ProjectModule } from '../project/project.module';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { UserProjectsComponent } from './user-projects/user-projects.component';
+import { ProjectModule } from 'src/app/modules/project/project.module';
+import { FormsModule } from '@angular/forms';
 
 
 @NgModule({
   declarations: [
-    UserProjectComponent
+    UserProjectsComponent
   ],
   imports: [
     CommonModule,
     UserRoutingModule,
-    PaginationModule.forRoot(),
     ProjectModule,
-    FormsModule,
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
+    FormsModule
   ]
 })
-export class UserModule { }
+export class UserModule {
+}
